+++ lib/flip_layout.dart
@@ -1,2 +1,3 @@
 export 'src/layout_motion.dart';
 export 'src/animated_layout.dart';
+export 'src/motion_group.dart';